Translate and fix last macro Moke stroke volume estimate macro work
